Applied SRP To DataManager Class
Extracted out the Hashing functionality from the DataManager class to a
new class specifically focused on encryption; PasswordManager. Because
encryption is not necessarily a part of the database connection, it
should not be in the DataManager class. By moving all encryption out of
the DataManager class and into its own, encryption-only class, it is
easier to expand and implement additional encryption functions without
muddying or confusing the DataManager class, as well as keeping it from
becoming bloated.